Add device entity, port to tokio
This ports the code to use tokio as the async framework for handling the
socket reading and message passing around, as well as introduces the
Device abstraction.

Moving to tokio seems to lead to cleaner code in terms of dealing with
reading/writing socket data as well as doing event loops on the
controller to handle device messages.

Speaking of devices, they are a new entity that encapsulates a
connection to a OF entity. They are responsible for keeping an
internal state of the device (only the device id for now) and they post
all received messages up to the controller through a mpsc channel.

The controller will get those messages and act on them. Right now it
does nothing, this is the next step to be done.